+++ src/sll/api/json_api.c
@@ -4,6 +4,7 @@
 #include <sll/api/math.h>
 #include <sll/array.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/file.h>
 #include <sll/map.h>
 #include <sll/memory.h>