+++ src/sll/api/path_api.c
@@ -3,6 +3,7 @@
 #include <sll/api/path.h>
 #include <sll/array.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/file.h>
 #include <sll/gc.h>
 #include <sll/memory.h>