+++ src/sll/api/string_api.c
@@ -2,6 +2,7 @@
 #include <sll/api.h>
 #include <sll/api/string.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/gc.h>
 #include <sll/memory.h>
 #include <sll/object.h>
@@ -9,7 +10,6 @@
 #include <sll/static_object.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <sll/util.h>
 #include <sll/var_arg.h>
 #include <sll/vm.h>
 #include <stdint.h>