+++ src/sll/api/sys_api.c
@@ -4,6 +4,7 @@
 #include <sll/api/sys.h>
 #include <sll/array.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/gc.h>
 #include <sll/map.h>
 #include <sll/memory.h>