+++ src/sll/platform/posix/posix_util.c
@@ -1,5 +1,6 @@
 #include <sll/_sll_internal.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/memory.h>
 #include <sll/string.h>
 #include <sll/types.h>