+++ src/sll/platform/windows/win_path.c
@@ -2,10 +2,10 @@
 #include <winioctl.h>
 #include <sll/_sll_internal.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/memory.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <sll/util.h>