+++ src/sll/platform/windows/win_util.c
@@ -2,12 +2,12 @@
 #include <bcrypt.h>
 #include <sll/_sll_internal.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/memory.h>
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/util.h>
 #include <signal.h>
-#include <stdint.h>