+++ src/sll/string.c
@@ -1,11 +1,12 @@
 #include <sll/_sll_internal.h>
 #include <sll/array.h>
 #include <sll/common.h>
+#include <sll/data.h>
 #include <sll/gc.h>
 #include <sll/map.h>
 #include <sll/memory.h>
-#include <sll/operator.h>
 #include <sll/object.h>
+#include <sll/operator.h>
 #include <sll/static_object.h>
 #include <sll/string.h>
 #include <sll/types.h>