+++ src/sll/string_format.c
@@ -3,7 +3,7 @@
 #include <sll/memory.h>
 #include <sll/string.h>
 #include <sll/types.h>
-#include <sll/util.h>
+#include <sll/data.h>
 #include <sll/var_arg.h>
 #include <stdarg.h>
 #include <stdint.h>