+++ src/sll/util.c
@@ -9,8 +9,6 @@
 #include <sll/string.h>
 #include <sll/types.h>
 #include <sll/util.h>
-#include <sll/version.h>
-#include <stdint.h>
 
 
 
@@ -20,22 +18,18 @@ STATIC_STRING_SETUP;
 
 static sll_cleanup_function_t _util_exit_table[MAX_CLEANUP_TABLE_SIZE];
 static sll_bool_t _util_init=0;
-static uint16_t _util_exit_table_size=0;
+static unsigned int _util_exit_table_size=0;
 static sll_sandbox_flags_t _util_sandbox_flags=0;
 
 
 
-__SLL_NO_RETURN void _force_exit(const sll_char_t* a,const sll_char_t* b,const sll_char_t* c,const sll_char_t* d,const sll_char_t* e){
+__SLL_NO_RETURN void _force_exit(const sll_char_t* a,const sll_char_t* b,const sll_char_t* c){
 	sll_file_flush(sll_stdout);
 	sll_file_flush(sll_stderr);
 	sll_file_descriptor_t fd=sll_platform_get_default_stream_descriptor(SLL_PLATFORM_STREAM_ERROR);
 	sll_platform_file_write(fd,a,sll_string_length_unaligned(a));
 	sll_platform_file_write(fd,b,sll_string_length_unaligned(b));
 	sll_platform_file_write(fd,c,sll_string_length_unaligned(c));
-	if (d){
-		sll_platform_file_write(fd,d,sll_string_length_unaligned(d));
-		sll_platform_file_write(fd,e,sll_string_length_unaligned(e));
-	}
 	_force_exit_platform();
 }
 
@@ -70,142 +64,6 @@ __SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_string_index_t sll_add_string_object(sll_s
 	v->t|=OBJECT_EXTERNAL_STRING;
 	*(st->dt+st->l-1)=v->dt.s;
 	return st->l-1;
-}
-
-
-
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_compare_result_t sll_compare_data(const void* a,const void* b,sll_size_t l){
-	if (!l){
-		return SLL_COMPARE_RESULT_EQUAL;
-	}
-	const sll_char_t* ap=(const sll_char_t*)a;
-	const sll_char_t* bp=(const sll_char_t*)b;
-	if (l<16){
-		for (sll_size_t i=0;i<l;i++){
-			if (*(ap+i)!=*(bp+i)){
-				return (*(ap+i)<*(bp+i)?SLL_COMPARE_RESULT_BELOW:SLL_COMPARE_RESULT_ABOVE);
-			}
-		}
-		return SLL_COMPARE_RESULT_EQUAL;
-	}
-	if ((((uint64_t)ap)&7)&&(((uint64_t)bp)&7)){
-		if ((((uint64_t)ap)&7)<(((uint64_t)bp)&7)){
-			ap=(const sll_char_t*)b;
-			bp=(const sll_char_t*)a;
-		}
-		SLL_UNIMPLEMENTED();
-	}
-	else if (((uint64_t)ap)&7){
-		ap=(const sll_char_t*)b;
-		bp=(const sll_char_t*)a;
-	}
-	const uint64_t* ap64=(const uint64_t*)ap;
-	const uint64_t* bp64=(const uint64_t*)bp;
-	SLL_ASSERT(!(((uint64_t)ap64)&7));
-	do{
-		if (*ap64!=*bp64){
-			uint64_t av=ROTATE_BITS64(*ap64,32);
-			uint64_t bv=ROTATE_BITS64(*bp64,32);
-			av=((av&0xffff0000ffffull)<<16)|((av&0xffff0000ffff0000ull)>>16);
-			bv=((bv&0xffff0000ffffull)<<16)|((bv&0xffff0000ffff0000ull)>>16);
-			return ((((av&0xff00ff00ff00ffull)<<8)|((av&0xff00ff00ff00ff00ull)>>8))<(((bv&0xff00ff00ff00ffull)<<8)|((bv&0xff00ff00ff00ff00ull)>>8))?SLL_COMPARE_RESULT_BELOW:SLL_COMPARE_RESULT_ABOVE);
-		}
-		ap64++;
-		bp64++;
-		l-=8;
-	} while (l>7);
-	ap=(const sll_char_t*)ap64;
-	bp=(const sll_char_t*)bp64;
-	while (l){
-		l--;
-		if (*ap!=*bp){
-			return (*ap<*bp?SLL_COMPARE_RESULT_BELOW:SLL_COMPARE_RESULT_ABOVE);
-		}
-		ap++;
-		bp++;
-	}
-	return SLL_COMPARE_RESULT_EQUAL;
-}
-
-
-
-__SLL_EXTERNAL void sll_copy_data(const void* s,sll_size_t l,void* d){
-	const sll_char_t* a=(const sll_char_t*)s;
-	sll_char_t* b=(sll_char_t*)d;
-	if (l<16){
-		while (l){
-			*b=*a;
-			a++;
-			b++;
-			l--;
-		}
-		return;
-	}
-	if (((uint64_t)d)&7){
-		sll_size_t i=8-(((uint64_t)d)&7);
-		a+=i;
-		b+=i;
-		l-=i;
-		do{
-			*(b-i)=*(a-i);
-			i--;
-		} while (i);
-	}
-	SLL_ASSERT(!(((uint64_t)b)&7));
-	const uint64_t* ap=(const uint64_t*)a;
-	uint64_t* bp=(uint64_t*)b;
-	ASSUME_ALIGNED(bp,3,0);
-	sll_size_t i=0;
-	for (;i<(l>>3);i++){
-		*(bp+i)=*(ap+i);
-	}
-	if (l&7){
-		l=(l&7)<<3;
-		*(bp+i)=((*(bp+i))&(0xffffffffffffffffull<<l))|((*(ap+i))&((1ull<<l)-1));
-	}
-}
-
-
-
-__SLL_EXTERNAL void* sll_copy_string(const sll_char_t* s,void* d){
-	if (!(*s)){
-		return d;
-	}
-	sll_char_t* o=(sll_char_t*)d;
-	while (((uint64_t)o)&7){
-		*o=*s;
-		s++;
-		o++;
-		if (!(*s)){
-			return o;
-		}
-	}
-	SLL_ASSERT(!(((uint64_t)o)&7));
-	const uint64_t* sp=(const uint64_t*)s;
-	uint64_t* op=(uint64_t*)o;
-	ASSUME_ALIGNED(op,3,0);
-	while (1){
-		uint64_t v=((*sp)-0x101010101010101ull)&0x8080808080808080ull&(~(*sp));
-		if (v){
-			SLL_ASSERT(FIND_FIRST_SET_BIT(v)>6&&(FIND_FIRST_SET_BIT(v)&7)==7);
-			v=FIND_FIRST_SET_BIT(v);
-			o=(void*)(((uint64_t)op)+(v>>3));
-			if (v>7){
-				v-=7;
-				*op=((*op)&(0xffffffffffffffffull<<v))|((*sp)&((1ull<<v)-1));
-			}
-			return o;
-		}
-		*op=*sp;
-		sp++;
-		op++;
-	}
-}
-
-
-
-__SLL_EXTERNAL void sll_copy_string_null(const sll_char_t* s,void* d){
-	*((sll_char_t*)sll_copy_string(s,d))=0;
 }
 
 
@@ -383,81 +241,7 @@ __SLL_EXTERNAL void sll_register_cleanup(sll_cleanup_function_t f){
 
 
 
-__SLL_EXTERNAL void sll_set_memory(void* p,sll_size_t l,sll_char_t v){
-	sll_char_t* o=(sll_char_t*)p;
-	if (l<16){
-		while (l){
-			*o=v;
-			o++;
-			l--;
-		}
-		return;
-	}
-	if (((uint64_t)o)&7){
-		sll_size_t i=8-(((uint64_t)o)&7);
-		l-=i;
-		do{
-			*o=v;
-			o++;
-			i--;
-		} while (i);
-	}
-	SLL_ASSERT(!(((uint64_t)o)&7));
-	uint64_t* op=(uint64_t*)o;
-	ASSUME_ALIGNED(op,3,0);
-	sll_size_t i=0;
-	uint64_t v64=0x101010101010101ull*v;
-	for (;i<(l>>3);i++){
-		*(op+i)=v64;
-	}
-	if (l&7){
-		l=(l&7)<<3;
-		*(op+i)=((*(op+i))&(0xffffffffffffffffull<<l))|(v64&((1ull<<l)-1));
-	}
-}
-
-
-
 __SLL_EXTERNAL sll_sandbox_flags_t sll_set_sandbox_flags(sll_sandbox_flags_t f){
 	_util_sandbox_flags|=f;
 	return _util_sandbox_flags;
-}
-
-
-
-__SLL_EXTERNAL __SLL_CHECK_OUTPUT sll_version_t sll_version(void){
-	return SLL_VERSION;
-}
-
-
-
-__SLL_EXTERNAL void sll_zero_memory(void* p,sll_size_t l){
-	sll_char_t* o=(sll_char_t*)p;
-	if (l<16){
-		while (l){
-			*o=0;
-			o++;
-			l--;
-		}
-		return;
-	}
-	if (((uint64_t)o)&7){
-		sll_size_t i=8-(((uint64_t)o)&7);
-		l-=i;
-		do{
-			*o=0;
-			o++;
-			i--;
-		} while (i);
-	}
-	SLL_ASSERT(!(((uint64_t)o)&7));
-	uint64_t* op=(uint64_t*)o;
-	ASSUME_ALIGNED(op,3,0);
-	sll_size_t i=0;
-	for (;i<(l>>3);i++){
-		*(op+i)=0;
-	}
-	if (l&7){
-		*(op+i)&=0xffffffffffffffffull<<((l&7)<<3);
-	}
 }