+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 
 ### Changed
 
+- Moved `SLL_COPY_STRING_NULL`, `sll_compare_data`, `sll_copy_data`, `sll_copy_string`, `sll_copy_string_null`, `sll_set_memory` and `sll_zero_memory` from [`src/sll/util.c`][0.6.37/src/sll/util.c] and [`src/sll/include/sll/util.h`][0.6.37/src/sll/include/sll/util.h] to [`src/sll/data.c`][0.6.38/src/sll/data.c] and [`src/sll/include/sll/data.h`][0.6.38/src/sll/include/sll/data.h]
+- Moved `sll_version` from [`src/sll/util.c`][0.6.37/src/sll/util.c] and [`src/sll/include/sll/util.h`][0.6.37/src/sll/include/sll/util.h] to [`src/sll/version.c`][0.6.38/src/sll/version.c] and [`src/sll/include/sll/version.h`][0.6.38/src/sll/include/sll/version.h]
 - Renamed `sll_api_time_current_nanos`, `sll_api_time_sleep_nanos`, `time$time_nanos` and `time$sleep_nanos` to `sll_api_time_current_ns`, `sll_api_time_sleep_ns`, `time$time_ns` and `time$sleep_ns`
 - Renamed [`src/sll/api/_generated.c`][0.6.37/src/sll/api/_generated.c] and [`src/sll/include/sll/api/_generated.h`][0.6.37/src/sll/include/sll/api/_generated.h] to [`src/sll/api/_generated_raw.c`][0.6.38/src/sll/api/_generated_raw.c] and [`src/sll/include/sll/api/_generated_raw.h`][0.6.38/src/sll/include/sll/api/_generated_raw.h]
 - `sll_platform_string` is now a `const sll_string_t*` instead of `const sll_char_t*`
@@ -1623,9 +1625,15 @@ Unfortunately, no versions were assigned before 2021-6-15 (:disappointed:), so t
 [#2]: https://github.com/sl-lang/sll/issues/2
 [test-coverage]: https://github.com/sl-lang/sll/tree/test-coverage
 [0.6.38/src/sll/api/_generated_raw.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.38/src/sll/api/_generated_raw.c
+[0.6.38/src/sll/data.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.38/src/sll/data.c
 [0.6.38/src/sll/include/sll/api/_generated_raw.h]: https://github.com/sl-lang/sll/blob/sll-v0.6.38/src/sll/include/sll/api/_generated_raw.h
+[0.6.38/src/sll/include/sll/data.h]: https://github.com/sl-lang/sll/blob/sll-v0.6.38/src/sll/include/sll/data.h
+[0.6.38/src/sll/include/sll/version.h]: https://github.com/sl-lang/sll/blob/sll-v0.6.38/src/sll/include/sll/version.h
+[0.6.38/src/sll/version.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.38/src/sll/version.c
 [0.6.37/src/sll/api/_generated.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.37/src/sll/api/_generated.c
 [0.6.37/src/sll/include/sll/api/_generated.h]: https://github.com/sl-lang/sll/blob/sll-v0.6.37/src/sll/include/sll/api/_generated.h
+[0.6.37/src/sll/include/sll/util.h]: https://github.com/sl-lang/sll/blob/sll-v0.6.37/src/sll/include/sll/util.h
+[0.6.37/src/sll/util.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.37/src/sll/util.c
 [0.6.36/src/sll/cli.c]: https://github.com/sl-lang/sll/blob/main/src/sll/cli.c
 [0.6.35/src/cli/main.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.35/src/cli/main.c
 [0.6.35/src/sll/cast.c]: https://github.com/sl-lang/sll/blob/sll-v0.6.35/src/sll/cast.c